+++ includes/class-wc-heidelpay-response.php
@@ -48,9 +48,12 @@ class WC_Heidelpay_Response
             self::$response->verifySecurityHash($secret, $identificationTransactionId);
         } catch (\Exception $exception) {
             $callers = debug_backtrace();
-            wc_get_logger()->log(WC_Log_Levels::NOTICE, print_r("Heidelpay - " .
-                $callers [0] ['function'] . ": Invalid response hash from " .
-                $_SERVER ['REMOTE_ADDR'] . ", suspecting manipulation", 1));
+            wc_get_logger()->notice(
+                print_r("Heidelpay - " .
+                    $callers [0] ['function'] . ": Invalid response hash from " .
+                    $_SERVER ['REMOTE_ADDR'] . ", suspecting manipulation", 1),
+                ['source' => 'heidelpay']
+            );
             exit(); //error
         }
 
@@ -58,41 +61,65 @@ class WC_Heidelpay_Response
         $order = wc_get_order($orderId);
 
 
-        $this->handleResult($post_data, $order);
+        $this->handleResult(self::$response, $order);
     }
 
     /**
      * handle result post
      *
-     * @param $post_data
+     * @param Response $response
      * @param WC_Order $order
      */
-    public function handleResult($post_data, WC_Order $order)
+    public function handleResult($response, WC_Order $order)
     {
-        $uid = self::$response->getIdentification()->getUniqueId();
-        $sid = self::$response->getIdentification()->getShortId();
+        $uid = $response->getIdentification()->getUniqueId();
+        $sid = $response->getIdentification()->getShortId();
+        $payCode = explode('.', strtoupper($response->getPayment()->getCode()));
+        list($transactionMethod, $transactionType) = $payCode;
 
-        if (self::$response->isSuccess()) {
-            $payCode = explode('.', strtoupper($post_data['PAYMENT_CODE']));
+        // Get Payment Method.
+        $paymentGatewayList = WC_Payment_Gateways::instance()->payment_gateways();
+        $paymentMethodId = $order->get_payment_method();
+        /** @var WC_Heidelpay_Payment_Gateway $paymentMethod */
+        $paymentMethod = !empty($paymentGatewayList[$paymentMethodId]) ? $paymentGatewayList[$paymentMethodId] : null;
+        if (!$paymentMethod || !($paymentMethod instanceof WC_Heidelpay_Payment_Gateway)) {
+            wc_get_logger()->notice(
+                sprintf("Payment method is not valid or was not found: %s", htmlspecialchars($paymentMethodId)),
+                ['source' => 'heidelpay']
+            );
+            return;
+        }
+
+        // If registration, do a debit on registration afterwards
+        if (($transactionType === 'RG' || $transactionType === 'CF') && $response->isSuccess()) {
+            $order->add_meta_data('heidelpay-Registration', $uid);
+            $order->save_meta_data();
+            /** @var WC_Heidelpay_Payment_Gateway $paymethod */
+            $paymentMethod->prepareRequest($order);
+            //$paymentMethod->payMethod->getRequest()->getFrontend()->setEnabled('FALSE');
+            $paymentMethod->payMethod->getRequest()->getIdentification()->setReferenceid($uid);
+            // Use the Response of the debitOnRegistration in order to set the correct paymentInfo
+            $debitOnRegistrationResponse = $paymentMethod->performNoGuiRequest($order, $uid);
+            if ($debitOnRegistrationResponse !== null) {
+                $redirectUrl = $debitOnRegistrationResponse->getFrontend()->getRedirectUrl();
+                if (!empty($redirectUrl)) {
+                    echo $redirectUrl;
+                    return;
+                }
+                $response = $debitOnRegistrationResponse;
+            }
+        }
+
+        if ($response->isSuccess() && !$response->isPending()) {
             $note = '';
 
-            $this->setPaymentInfo($order);
-
-            // If registration, do a debit on registration afterwards
-            if ($payCode[1] === 'RG' || $payCode[1] === 'CF') {
-                $order->add_meta_data('heidelpay-Registration', $uid);
-                $paymethod = 'WC_Gateway_HP_' . $payCode[0];
-                $paymethod = new $paymethod;
-                $paymethod->prepareRequest($order);
-                $paymethod->payMethod->getRequest()->getFrontend()->setEnabled('FALSE');
-                $paymethod->payMethod->getRequest()->getIdentification()->setReferenceid($uid);
-                $paymethod->performNoGuiRequest($order, $uid);
-            }
+            $paymentMethod->setPaymentInfo($order, $response);
+            $order->add_meta_data('heidelpay-UniqueID', $uid);
+            $order->add_meta_data('heidelpay-ShortID', $sid);
 
             // If no money has been payed yet.
-            if ($payCode[0] !== 'IV' && $payCode[1] === 'PA') {
-                // If not Prepayment and Invoice payment can be captured manually
-                if ($payCode [0] !== 'PP') {
+            if ($transactionType === 'PA') {
+                if ($transactionMethod !== 'PP') {
                     $note = __(
                         'Payment reservation successful. Please use the hiP to check the payment.',
                         'woocommerce-heidelpay'
@@ -104,87 +131,27 @@ class WC_Heidelpay_Response
                     __('Awaiting payment.', 'woocommerce-heidelpay') . ' ' . $note
                 );
             } else {
-                $order->add_meta_data('heidelpay-UniqueID', $uid);
-                $order->add_meta_data('heidelpay-ShortID', $sid);
                 $order->payment_complete($sid);
             }
-            echo $order->get_checkout_order_received_url();
-
             /* redirect customer to success page */
-        } elseif (self::$response->isError()) {
-            $error = self::$response->getError();
+            echo $order->get_checkout_order_received_url();
+        } elseif ($response->isError()) {
+            $error = $response->getError();
             $order->update_status('failed');
 
-            echo apply_filters('woocommerce_get_cancel_order_url_raw', add_query_arg(array(
+            echo apply_filters('woocommerce_get_cancel_order_url_raw', add_query_arg([
                 'cancel_order' => 'true',
                 'order' => $order->get_order_key(),
                 'order_id' => $order->get_id(),
                 '_wpnonce' => wp_create_nonce('woocommerce-cancel_order'),
                 'errorCode' => $error['code'],
-            ), $order->get_cancel_endpoint()));
-        } elseif (self::$response->isPending()) {
+            ], $order->get_cancel_endpoint()));
+        } elseif ($response->isPending()) {
             //empty cart
             wc()->cart->empty_cart();
 
             //show thank you page
             echo $order->get_checkout_order_received_url();
-        }
-    }
-
-    /**
-     * Add payment information to the order if available.
-     * Information usually are set for invoice, direct debit and prepayment.
-     * @param WC_Order $order
-     * @return null
-     */
-    public function setPaymentInfo(WC_Order $order)
-    {
-        // Load template text for Payment information
-        $payInfoTemplate = $this->getInfoTemplate();
-        if ($payInfoTemplate === null) {
-            return null;
-        }
-
-        $response = self::$response;
-        $payInfo = $response->getConnector();
-        $presentation = $response->getPresentation();
-
-        $paymentData = array(
-            '{AMOUNT}' => $presentation->getAmount(),
-            '{CURRENCY}' => $presentation->getCurrency(),
-            '{CONNECTOR_ACCOUNT_HOLDER}' => $payInfo->getAccountHolder(),
-            '{CONNECTOR_ACCOUNT_IBAN}' => $payInfo->getAccountIBan(),
-            '{CONNECTOR_ACCOUNT_BIC}' => $payInfo->getAccountBic(),
-            '{IDENTIFICATION_SHORTID}' => self::$response->getIdentification()->getShortId(),
-            '{Iban}' => $response->getAccount()->getIban(),
-            '{Ident}' => $response->getAccount()->getIdentification(),
-            '{CreditorId}' => $response->getIdentification()->getCreditorId(),
-        );
-
-        $paymentText = strtr($payInfoTemplate, $paymentData);
-
-        $order->add_meta_data('heidelpay-paymentInfo', $paymentText);
-    }
-
-    /**
-     * Provide the template text for payment information.
-     * @return null|string
-     */
-    public function getInfoTemplate()
-    {
-        $payCode = explode('.', self::$response->getPayment()->getCode());
-        switch (strtoupper($payCode[0])) {
-            case 'IV':
-                return __('invoice_info', 'woocommerce-heidelpay');
-                break;
-            case 'PP':
-                return __('prepayment_info', 'woocommerce-heidelpay');
-                break;
-            case 'DD':
-                return __('direct_debit_info', 'woocommerce-heidelpay');
-                break;
-            default:
-                return null;
         }
     }
 }