+++ includes/gateways/class-wc-heidelpay-gateway-pp.php
@@ -49,10 +49,6 @@ class WC_Gateway_HP_PP extends WC_Heidelpay_Payment_Gateway
             __('Insert payment data for %s', 'woocommerce-heidelpay'),
             $this->name
         );
-        $this->form_fields['instructions']['default'] = __(
-            'please send the money to IBAN BIC ',
-            'woocommerce-heidelpay'
-        );
         $this->form_fields['enabled']['label'] = sprintf(__('Enable %s', 'woocommerce-heidelpay'), $this->name);
         $this->form_fields['security_sender']['default'] = '31HA07BC8142C5A171745D00AD63D182';
         $this->form_fields['user_login']['default'] = '31ha07bc8142c5a171744e5aef11ffd3';
@@ -82,5 +78,10 @@ class WC_Gateway_HP_PP extends WC_Heidelpay_Payment_Gateway
         $this->name = __('Prepayment', 'woocommerce-heidelpay');
         $this->has_fields = false;
         $this->bookingAction = 'authorize';
+    }
+
+    public function getPaymentInfoTemplate()
+    {
+        return __('prepayment_info', 'woocommerce-heidelpay');
     }
 }